♻️ refactor: eventHandler 리팩토링
