+++ Event/Events.js
@@ -1,0 +1,9 @@
+import { historyTabEventHandler } from './HistoryEvent.js';
+import { FabEventHandler } from './FabEvent.js';
+import { modalEventHandler } from './ModalEvent.js';
+
+export const EventHandler = (state) => {
+  historyTabEventHandler();
+  FabEventHandler(state);
+  modalEventHandler(state);
+};
