+++ Event/FabEvent.js
@@ -1,0 +1,10 @@
+import { $ } from '../util.js';
+import { addNewColumn } from '../View/ColumnView.js';
+
+const btnFAB = $('.fab');
+
+export const FabEventHandler = (state) => {
+  btnFAB.addEventListener('click', () => {
+    addNewColumn(state);
+  });
+};
